% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoFTestCpp.R
\name{goftest_cpp}
\alias{goftest_cpp}
\title{Monte Carlo Goodness of Fit tests for Stochastic Block Models}
\usage{
goftest_cpp(A, K, numGraphs = 100)
}
\arguments{
\item{A}{a n by n binary symmetric adjacency matrix representing a undirected graph where n is the no nodes in the graph}

\item{K}{a numeric scalar representing no of blocks}

\item{numGraphs}{number of graphs will be sampled; default value is 100}
}
\value{
A list with the elements
\item{statistic}{Value of the chi-square test statistic}
\item{p.value}{the p-value for the test.}
}
\description{
\code{goftest_cpp} performs chi square goodness of fit test for network data considering the model as ERSBM using  Rcpp
}
